corrected checkstyle violations
